Изменила ЛР4 для ЛР5 и добавила пользователей
Изменила ЛР4 для ЛР5 и добавила пользователей

+++ LB4/delete_games.php
@@ -1,8 +1,13 @@
 <?php
-$mysqli = new mysqli("localhost", "a0609677_root", "root", "a0609677_games") or die ("Невозможно
+include("checks.php");
+require_once 'connect1.php';
+$mysqli = new mysqli($host, $user, $password, $database) or die ("Невозможно
 подключиться к серверу");
 $id_games = $_GET['id_games'];
-$result = $mysqli->query("DELETE FROM games WHERE id_games='$id_games'");
-header("Location: index.php");
+if ($_SESSION['type'] == 2)
+    $result = $mysqli->query("DELETE FROM games WHERE id_games ='$id_games'");
+else
+    echo "Недостаточно прав";
+header("Location: gamesAdm.php");
 exit;
 ?>
