+++ LB4/delete_key.php
@@ -1,8 +1,13 @@
 <?php
-$mysqli = new mysqli("localhost", "a0609677_root", "root", "a0609677_games") or die ("Невозможно
+include("checks.php");
+require_once 'connect1.php';
+$mysqli = new mysqli($host, $user, $password, $database) or die ("Невозможно
 подключиться к серверу");
 $id_key = $_GET['id_key'];
-$result = $mysqli->query("DELETE FROM kl WHERE id_key='$id_key'");
-header("Location: key.php");
+if ($_SESSION['type'] == 2)
+    $result = $mysqli->query("DELETE FROM kl WHERE id_key='$id_key'");
+else
+    echo "Недостаточно прав";
+header("Location: keyAdm.php");
 exit;
 ?>
