+++ LB4/delete_store.php
@@ -1,8 +1,13 @@
 <?php
-$mysqli = new mysqli("localhost", "a0609677_root", "root", "a0609677_games") or die ("Невозможно
-подключиться к серверу");
+include("checks.php");
+require_once 'connect1.php';
+$mysqli = new mysqli($host, $user, $password, $database)
+or die ("Невозможно подключиться к серверу");
 $id_stores = $_GET['id_stores'];
-$result = $mysqli->query("DELETE FROM stores WHERE id_stores='$id_stores'");
-header("Location: stores.php");
+if ($_SESSION['type'] == 2)
+    $result = $mysqli->query("DELETE FROM stores WHERE id_stores='$id_stores'");
+else
+    echo "Недостаточно прав";
+header("Location: storesAdm.php");
 exit;
 ?>
